refactor(sql): split oversized planner and resolver modules
Break the two largest files in the SQL crate into directories, one module
per concern. No behavior change.

- resolver/expr/convert.rs becomes convert/, split by expression family:
  entry, identifier, operators, predicates, literals, builtins
- planner/subquery.rs becomes subquery/, split by subquery kind:
  extract, in_list, scalar
- Move array table-valued-function resolution to resolver/array_tvf.rs
- Move the comma-LATERAL branch to planner/select/comma_lateral.rs
- Move CteCatalog to planner/select/cte_catalog.rs
- Move LATERAL AST extraction to planner/lateral/subquery.rs
