+++ nodedb-sql/src/planner/select/mod.rs
@@ -6,6 +6,8 @@
 //! search patterns (vector, text, hybrid, spatial) directly from the AST
 //! instead of reverse-engineering an optimizer's output.
 
+mod comma_lateral;
+mod cte_catalog;
 mod derived_from;
 mod entry;
 mod entry_ann;
@@ -17,8 +19,10 @@ mod query_tail;
 mod select_stmt;
 mod where_search;
 
+pub(crate) use cte_catalog::CteCatalog;
 pub use entry::plan_query;
 pub use helpers::{
     convert_projection, convert_where_to_filters, extract_float, extract_func_args,
     extract_string_literal, qualified_name,
 };
+pub(crate) use order_by::select_output_aliases;
