+++ nodedb-sql/src/resolver/expr/convert/mod.rs
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! Convert sqlparser AST expressions to our SqlExpr IR.
+
+pub mod builtins;
+pub mod entry;
+pub mod identifier;
+pub mod literals;
+pub mod operators;
+pub mod predicates;
+
+pub use entry::convert_expr;
+pub(in crate::resolver::expr) use entry::convert_expr_depth;
